fixes suggested by coderabbit
