+++ packages/rs-dpp/src/tokens/errors.rs
@@ -6,6 +6,6 @@ pub enum TokenError {
     TokenNotFoundAtPositionError,
     #[error("The contract version does not allow tokens")]
     TokenNotFoundOnContractVersion,
-    #[error("There is minting recipient set")]
+    #[error("There is no minting recipient set")]
     TokenNoMintingRecipient,
 }
