+++ packages/rs-drive/src/drive/tokens/info/queries.rs
@@ -1,5 +1,5 @@
 use crate::drive::tokens::paths::{
-    token_identity_infos_path_vec, token_identity_infos_root_path_vec, token_statuses_root_path_vec,
+    token_identity_infos_path_vec, token_identity_infos_root_path_vec,
 };
 use crate::drive::Drive;
 use crate::query::{Query, QueryItem};
@@ -53,33 +53,6 @@ impl Drive {
             tokens_root,
             SizedQuery::new(query, Some(token_ids.len() as u16), None),
         )
-    }
-
-    /// The query getting a token statuses
-    pub fn token_statuses_query(token_ids: &[[u8; 32]]) -> PathQuery {
-        let tokens_root = token_statuses_root_path_vec();
-
-        let mut query = Query::new();
-
-        for token_id in token_ids {
-            query.insert_key(token_id.to_vec());
-        }
-
-        PathQuery::new(
-            tokens_root,
-            SizedQuery::new(query, Some(token_ids.len() as u16), None),
-        )
-    }
-
-    /// The query getting a token statuses
-    pub fn token_status_query(token_id: [u8; 32]) -> PathQuery {
-        let tokens_root = token_statuses_root_path_vec();
-
-        let mut query = Query::new();
-
-        query.insert_key(token_id.to_vec());
-
-        PathQuery::new(tokens_root, SizedQuery::new(query, Some(1), None))
     }
 
     /// The query getting token infos for identities in a range
