+++ packages/rs-drive/src/drive/tokens/status/mod.rs
@@ -5,3 +5,4 @@ mod fetch_token_statuses;
 mod fetch_token_status;
 #[cfg(feature = "server")]
 mod prove_token_statuses;
+mod queries;
