+++ packages/rs-drive/src/verify/tokens/verify_token_balance_for_identity_id/mod.rs
@@ -12,35 +12,38 @@ use crate::verify::RootHash;
 use dpp::version::PlatformVersion;
 
 impl Drive {
-    /// Verifies the balance of a token held by a specific identity using a cryptographic proof.
+    /// Verifies the token balance of a specific identity using a cryptographic proof.
     ///
-    /// This method checks the cryptographic proof to verify the balances of a list of tokens
-    /// associated with the given identity ID. It dispatches to version-specific implementations
-    /// based on the platform version.
+    /// This function validates the token balance associated with an identity by verifying
+    /// the provided cryptographic proof. It ensures the correctness of the balance stored
+    /// for the given identity and token combination.
     ///
     /// # Parameters
-    /// - `proof`: The cryptographic proof to verify.
-    /// - `token_id`: A token ID to verify.
-    /// - `identity_id`: The unique identifier of the identity (32-byte array).
-    /// - `verify_subset_of_proof`: Whether to verify only a subset of the proof.
-    /// - `platform_version`: The current platform version.
+    ///
+    /// - `proof`: A slice of bytes containing the cryptographic proof of the token balance.
+    /// - `token_id`: A 32-byte identifier representing the unique ID of the token to verify.
+    /// - `identity_id`: A 32-byte identifier representing the identity whose token balance
+    ///   is to be verified.
+    /// - `verify_subset_of_proof`: A boolean indicating whether to verify only a subset of
+    ///   the provided proof.
+    /// - `platform_version`: A reference to the [PlatformVersion] object specifying which
+    ///   implementation version of the function to invoke.
     ///
     /// # Returns
-    /// - `Ok((RootHash, T))`:
-    ///   - `RootHash`: The verified root hash of the database.
-    ///   - `T`: A collection of `(token ID, token balance)` pairs.
+    ///
+    /// Returns a `Result` containing:
+    /// - `Ok((RootHash, Option<TokenAmount>))`: A tuple where:
+    ///   - `RootHash`: The root hash of the data structure at the time the proof was generated.
+    ///   - `Option<TokenAmount>`: The token balance if it exists, or `None` if the balance is absent.
+    /// - `Err(Error)`: An error if the verification fails due to an invalid proof, incorrect data,
+    ///   or version mismatch.
     ///
     /// # Errors
-    /// - `Error::Drive(DriveError::UnknownVersionMismatch)`:
-    ///   - Occurs when the platform version does not match any known version for this method.
-    /// - `Error::Proof(ProofError::WrongElementCount)`:
-    ///   - If the number of elements in the proof does not match the number of token IDs.
-    /// - `Error::Proof(ProofError::IncorrectValueSize)`:
-    ///   - If the token ID size or proof value size is invalid.
-    /// - `Error::Proof(ProofError::InvalidSumItemValue)`:
-    ///   - If the proof element does not represent a valid sum item.
-    /// - `Error::Proof(ProofError::InvalidItemType)`:
-    ///   - If the proof element is not a sum item as expected for balances.
+    ///
+    /// This function may return an `Error` in the following cases:
+    /// - The provided proof is invalid or corrupted.
+    /// - The token balance data is missing, inconsistent, or does not match the proof.
+    /// - The specified platform version does not match any known or supported implementations.
     pub fn verify_token_balance_for_identity_id(
         proof: &[u8],
         token_id: [u8; 32],
